Add global Snackbar event listener system
Adds static addEventListener()/removeEventListener() functions to the exported
Snackbar object.  Consuming modules can then add listeners so that they are
notified whenever a Snackbar is shown or hidden - for example, a custom Material
Design Floating Action Button component could add show and hide listeners for
moving itself out of the way of the Snackbar.

The available events are:
  - 'show': fired when the Snackbar is being shown
  - 'shown': fired once the Snackbar has finished animating and is fully visible
  - 'hide': fired as the Snackbar begins to hide
  - 'hidden': fired once the Snackbar has finished animating and is no longer
    visible

Example:

    const shownCallback = () => console.log('A snackbar was shown!')

    Snackbar.addEventListener('shown', shownCallback)

    // ...later, when we're finished:
    Snackbar.removeEventListener('shown', shownCallback)

We also export the durations of the Snackbar's show and hide animation
durations, so that consuming modules can time their animations to fit.
Unfortunately, these animation durations are all hardcoded and copied from
the Android source code as they're all private statics.

One final change is to use the module's dismiss() function in the Snackbar
action click handlers, so that the 'hide' event is sent if the Snackbar is
manually dismissed by the user.